+++ src/controllers/category.controller.ts
@@ -2,6 +2,7 @@ import { Response } from "express";
 import { IPaginationQuery, IReqUser } from "../utils/interfaces";
 import response from "../utils/response";
 import CategoryModel, { categoryDAO } from "../models/category.model";
+import { isValidObjectId } from "mongoose";
 
 export default {
   async create(req: IReqUser, res: Response) {
@@ -61,8 +62,11 @@ export default {
   async findOne(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
-      const result = await CategoryModel.findById(id);
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "category not found");
+      }
 
+      const result = await CategoryModel.findById(id);
       if (!result) {
         return response.notFound(res, "category not found");
       }
@@ -76,6 +80,10 @@ export default {
   async update(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "category not found");
+      }
+
       const result = await CategoryModel.findByIdAndUpdate(id, req.body, {
         new: true,
       });
@@ -88,6 +96,10 @@ export default {
   async remove(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "category not found");
+      }
+
       const result = await CategoryModel.findByIdAndDelete(id, { new: true });
       response.success(res, result, "success remove category");
     } catch (error) {
