+++ src/controllers/ticket.controller.ts
@@ -60,6 +60,10 @@ export default {
   async findOne(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "ticket not found");
+      }
+
       const result = await TicketModel.findById(id);
       if (!result) {
         return response.notFound(res, "ticket not found");
@@ -73,6 +77,10 @@ export default {
   async update(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "ticket not found");
+      }
+
       const result = await TicketModel.findByIdAndUpdate(id, req.body, {
         new: true,
       });
@@ -84,6 +92,10 @@ export default {
   async remove(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "ticket not found");
+      }
+
       const result = await TicketModel.findByIdAndDelete(id, {
         new: true,
       });
